requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
torch
ultralytics
python-dotenv
google-generativeai
Pillow
